pyproject.toml
[project]
name = "fastapi-react"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.114.2",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    # database
    "sqlalchemy[asyncio]>=2.0.41",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "python-jose>=3.5.0",
    "alembic>=1.16.2",
    "sqlalchemy-utils>=0.41.2",
    #cli
    "click>=8.2.1",
]

[dependency-groups]
dev = [
    {include-group = "test"},
    {include-group = "lint"},
]
test = [
    "pytest>=8.4.1",
]
lint = [
    "ruff>=0.12.5",
]

[project.scripts]
cli = "src.cli:main"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
exclude = ["alembic"]

